Refactor and Rename Storage and DataAdapter
