+++ src/main/java/storage/Storage.java
@@ -1,111 +1,125 @@
 package storage;
 
-import com.google.gson.JsonObject;
-import com.google.gson.JsonElement;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.google.gson.JsonParser;
-import java.util.logging.Logger;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import history.DateSerializer;
+import history.History;
+import programme.Day;
+import programme.ProgrammeList;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.LinkedHashMap;
+import java.util.Map;
 import java.util.logging.Level;
+import java.util.logging.Logger;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.io.FileReader;
-import java.io.IOException;
-
-/**
- * Represents the storage system for saving and loading tasks.
- * The <code>Storage</code> class handles reading from and writing to the file specified by the user.
- */
+/*
+    Data Manager acts as an adapter layer between the Storage and BuffBuddy classes,
+    translating between JSON and objects
+*/
 public class Storage {
-    private final String path;
-    private final Logger logger = Logger.getLogger(this.getClass().getName());
+
+    private static final Logger logger = Logger.getLogger(Storage.class.getName());
+    private final FileManager fileManager;
 
     public Storage(String path) {
-        this.path = path;
+        this.fileManager = new FileManager(path);
     }
 
-    public JsonObject loadProgrammeList() {
-        JsonObject jsonObject = load();
-        if(jsonObject == null || !jsonObject.has("programmeList")) {
-            logger.log(Level.INFO, "No programme list found.");
-            return new JsonObject();
-        }
-        logger.log(Level.INFO, "Programme list Loaded");
-        return jsonObject.getAsJsonObject("programmeList");
-    }
-
-    public JsonObject loadHistory() {
-        JsonObject jsonObject = load();
-        if(jsonObject == null || !jsonObject.has("history")) {
-            logger.log(Level.INFO, "No history found.");
-            return new JsonObject();
-        }
-        logger.log(Level.INFO, "History Loaded");
-        return jsonObject.getAsJsonObject("history");
-    }
-
-    private JsonObject load() {
-        logger.info("Attempting to load data from file: " + path);
-        try (FileReader reader = new FileReader(path)){
-            JsonElement element = JsonParser.parseReader(reader);
-            if(element == null || element.isJsonNull()) {
-                logger.info("No data found");
-                return new JsonObject();
-            }
-            logger.info("Data successfully loaded from file");
-            return element.getAsJsonObject();
-        } catch(IOException e){
-            logger.log(Level.WARNING, "Failed to load data from file: " + path, e);
-            throw new RuntimeException("Failed to load data due to: " + e.getMessage());
+    public ProgrammeList loadProgrammeList() {
+        try {
+            JsonObject programmeListJson = fileManager.loadProgrammeList();
+            return programmeListFromJson(programmeListJson);
+        } catch (Exception e) {
+            return new ProgrammeList();
         }
     }
 
-    public void save(JsonObject data) throws IOException {
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
-
-        createDirIfNotExists();
-        createFileIfNotExists();
-
-        try (FileWriter writer = new FileWriter(path)) {
-            gson.toJson(data, writer);
-            logger.info("Data successfully saved to file.");
-        } catch (IOException e) {
-            logger.log(Level.WARNING, "Failed to save data to file: " + path, e);
-            throw new IOException("Failed to save data due to: " + e.getMessage());
+    public History loadHistory() {
+        try {
+            JsonObject historyJson = fileManager.loadHistory();
+            return historyFromJson(historyJson);
+        } catch (Exception e) {
+            return new History();
         }
     }
 
-    private void createDirIfNotExists() throws IOException {
-        File dir = new File(path).getParentFile();
+    public void saveData(ProgrammeList pList, History history) {
+        assert pList != null : "programmeList must not be null";
+        assert history != null : "history must not be null";
 
-        if (dir == null || dir.exists()){
-            logger.log(Level.INFO, "Directory exists");
-            return;
+        JsonObject jsonObject = createJSON(pList, history);
+        logger.info("JsonObject containing programme list and history created.");
+
+        try{
+            fileManager.save(jsonObject);
+        } catch (Exception ignored) {
+            // For now, leave this as a quiet failure for simplicity
+            // User will be notified of corrupted data when next loading app
         }
-
-        boolean isSuccess = dir.mkdirs();
-
-        if (!isSuccess){
-            logger.log(Level.WARNING, "Failed to create directory.");
-            throw new IOException("Failed to create directory: " + dir.getAbsolutePath());
-        }
-        logger.log(Level.INFO, "Directory created");
     }
 
-    private void createFileIfNotExists() throws IOException {
-        File file = new File(path);
-        if (file.exists()) {
-            logger.log(Level.INFO, "File exists");
-            return;
-        }
+    private JsonObject createJSON(ProgrammeList programmeList, History history) {
+        JsonObject jsonObject = new JsonObject();
 
-        boolean isSuccess = file.createNewFile();
+        assert programmeList != null : "programmeList must not be null";
+        assert history != null : "history must not be null";
 
-        if (!isSuccess) {
-            logger.log(Level.WARNING, "Failed to create file.");
-            throw new IOException("Failed to create file: " + file.getAbsolutePath());
+        jsonObject.add("programmeList", programmeListToJson(programmeList));
+        logger.info("Programme list converted to JsonObject.");
+        jsonObject.add("history", historyToJson(history));
+        logger.info("History converted to JsonObject.");
+        return jsonObject;
+    }
+
+    private JsonObject programmeListToJson(ProgrammeList programmeList) {
+        Gson gson = new Gson();
+        logger.log(Level.INFO, "Programme list converted to Json for saving.");
+        return gson.toJsonTree(programmeList).getAsJsonObject();
+    }
+
+    private static ProgrammeList programmeListFromJson(JsonObject jsonObject) {
+        Gson gson = new Gson();
+        logger.log(Level.INFO, "Programme list converted from Json for loading.");
+        return gson.fromJson(jsonObject, ProgrammeList.class);
+    }
+
+    private JsonObject historyToJson(History history) {
+        Gson gson = new GsonBuilder()
+                .registerTypeAdapter(LocalDate.class, new DateSerializer())  // Custom serializer for LocalDate
+                .setPrettyPrinting()
+                .create();
+
+        JsonObject historyJson = new JsonObject();
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
+        LinkedHashMap<LocalDate, Day> historyMap = history.getHistory(); //To access the Hashmap
+
+        for (LocalDate date : historyMap.keySet()) {
+            Day day = historyMap.get(date);
+            historyJson.add(date.format(formatter), gson.toJsonTree(day));
         }
-        logger.log(Level.INFO, "File created");
+        logger.log(Level.INFO, "History converted to Json for saving.");
+        return historyJson;
+    }
+
+    private static History historyFromJson(JsonObject jsonObject) {
+        Gson gson = new GsonBuilder()
+                .registerTypeAdapter(LocalDate.class, new DateSerializer())  // Custom deserializer for LocalDate
+                .create();
+        History history = new History();
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
+        LinkedHashMap<LocalDate, Day> historyMap = history.getHistory(); //To access the Hashmap
+
+
+        for (Map.Entry<String, JsonElement> entry : jsonObject.entrySet()) {
+            LocalDate date = LocalDate.parse(entry.getKey(), formatter);
+            Day day = gson.fromJson(entry.getValue(), Day.class);
+            historyMap.put(date, day);
+        }
+        logger.log(Level.INFO, "historyJson converted from Json for loading.");
+        return history;
     }
 }
